pyproject.toml
[tool.poetry]
name = "nlp-with-llms"
version = "0.1.0"
description = ""
authors = ["shaan <skhoslajr@gmail.com>"]
readme = "README.md"
packages = [{include = "nlp_with_llms"}]

[tool.poetry.dependencies]
python = "^3.10"
nvidia-ml-py3 = "^7.352.0"
pytorch-lightning = "^2.0.1.post0"
transformers = "^4.28.0"
torchvision = "^0.15.1"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
